refactor: rename node to npm
