+++ pkg/detector/library/npm/compare.go
@@ -1,4 +1,4 @@
-package node
+package npm
 
 import (
 	"golang.org/x/xerrors"
@@ -8,16 +8,16 @@ import (
 	"github.com/aquasecurity/trivy/pkg/detector/library/comparer"
 )
 
-// NpmComparer represents a comparer for npm
-type NpmComparer struct{}
+// Comparer represents a comparer for npm
+type Comparer struct{}
 
 // IsVulnerable checks if the package version is vulnerable to the advisory.
-func (n NpmComparer) IsVulnerable(ver string, advisory dbTypes.Advisory) bool {
+func (n Comparer) IsVulnerable(ver string, advisory dbTypes.Advisory) bool {
 	return comparer.IsVulnerable(ver, advisory, n.matchVersion)
 }
 
 // matchVersion checks if the package version satisfies the given constraint.
-func (n NpmComparer) matchVersion(currentVersion, constraint string) (bool, error) {
+func (n Comparer) matchVersion(currentVersion, constraint string) (bool, error) {
 	v, err := npm.NewVersion(currentVersion)
 	if err != nil {
 		return false, xerrors.Errorf("npm version error (%s): %s", currentVersion, err)
