+++ pkg/detector/library/npm/testdata/fixtures/invalid-type.yaml
