+++ pkg/detector/library/npm/testdata/fixtures/no-value.yaml
