+++ pkg/detector/library/npm/testdata/fixtures/npm.yaml
